refactor: move From<v1/v2> impls from domain models to client modules
