+++ src/airflow/client/v1/dag.rs
@@ -3,7 +3,8 @@ use async_trait::async_trait;
 use log::{debug, info};
 use reqwest::Method;
 
-use crate::airflow::{model::common::DagList, traits::DagOperations};
+use crate::airflow::model::common::{Dag, DagList, Tag};
+use crate::airflow::traits::DagOperations;
 
 use super::model::dag::DagCollectionResponse;
 
@@ -68,7 +69,7 @@ impl DagOperations for V1Client {
         Ok(())
     }
 
-    async fn get_dag_code(&self, dag: &crate::airflow::model::common::Dag) -> Result<String> {
+    async fn get_dag_code(&self, dag: &Dag) -> Result<String> {
         let r = self
             .base_api(Method::GET, &format!("dagSources/{}", dag.file_token))
             .await?
@@ -76,5 +77,57 @@ impl DagOperations for V1Client {
         let response = self.base.client.execute(r).await?.error_for_status()?;
         let code = response.text().await?;
         Ok(code)
+    }
+}
+
+// From trait implementations for v1 models
+impl From<super::model::dag::DagResponse> for Dag {
+    fn from(value: super::model::dag::DagResponse) -> Self {
+        Self {
+            dag_id: value.dag_id.into(),
+            dag_display_name: Some(value.dag_display_name),
+            description: value.description,
+            fileloc: value.fileloc,
+            is_paused: value.is_paused.unwrap_or(false),
+            is_active: value.is_active,
+            has_import_errors: value.has_import_errors.unwrap_or(false),
+            has_task_concurrency_limits: value.has_task_concurrency_limits.unwrap_or(false),
+            last_parsed_time: value.last_parsed_time,
+            last_expired: value.last_expired,
+            max_active_tasks: value.max_active_tasks.unwrap_or(0),
+            max_active_runs: value.max_active_runs,
+            next_dagrun_logical_date: value.next_dagrun,
+            next_dagrun_data_interval_start: value.next_dagrun_data_interval_start,
+            next_dagrun_data_interval_end: value.next_dagrun_data_interval_end,
+            next_dagrun_create_after: value.next_dagrun_create_after,
+            owners: value.owners.clone(),
+            tags: value
+                .tags
+                .unwrap_or_default()
+                .into_iter()
+                .map(std::convert::Into::into)
+                .collect(),
+            file_token: value.file_token.clone(),
+            timetable_description: value.timetable_description,
+        }
+    }
+}
+
+impl From<DagCollectionResponse> for DagList {
+    fn from(value: DagCollectionResponse) -> Self {
+        Self {
+            dags: value
+                .dags
+                .into_iter()
+                .map(std::convert::Into::into)
+                .collect(),
+            total_entries: value.total_entries,
+        }
+    }
+}
+
+impl From<super::model::dag::DagTagResponse> for Tag {
+    fn from(value: super::model::dag::DagTagResponse) -> Self {
+        Self { name: value.name }
     }
 }
