+++ src/airflow/client/v1/dagstats.rs
@@ -3,7 +3,10 @@ use async_trait::async_trait;
 use reqwest::Method;
 
 use super::model;
-use crate::airflow::{model::common::DagStatsResponse, traits::DagStatsOperations};
+use crate::airflow::model::common::dagrun::DagRunState;
+use crate::airflow::model::common::dagstats::{DagStatistic, DagStatistics};
+use crate::airflow::model::common::DagStatsResponse;
+use crate::airflow::traits::DagStatsOperations;
 
 use super::V1Client;
 
@@ -20,5 +23,41 @@ impl DagStatsOperations for V1Client {
 
         let dag_stats = response.json::<model::dagstats::DagStatsResponse>().await?;
         Ok(dag_stats.into())
+    }
+}
+
+// From trait implementations for v1 models
+impl From<model::dagstats::DagStatsResponse> for DagStatsResponse {
+    fn from(value: model::dagstats::DagStatsResponse) -> Self {
+        Self {
+            dags: value
+                .dags
+                .into_iter()
+                .map(std::convert::Into::into)
+                .collect(),
+            total_entries: value.total_entries,
+        }
+    }
+}
+
+impl From<model::dagstats::DagStatistics> for DagStatistics {
+    fn from(value: model::dagstats::DagStatistics) -> Self {
+        Self {
+            dag_id: value.dag_id,
+            stats: value
+                .stats
+                .into_iter()
+                .map(std::convert::Into::into)
+                .collect(),
+        }
+    }
+}
+
+impl From<model::dagstats::DagStatistic> for DagStatistic {
+    fn from(value: model::dagstats::DagStatistic) -> Self {
+        Self {
+            state: DagRunState::from(value.state.as_str()),
+            count: value.count,
+        }
     }
 }
