+++ src/airflow/model/common/dag.rs
@@ -1,5 +1,3 @@
-use crate::airflow::client::v1;
-use crate::airflow::client::v2;
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
 
@@ -40,107 +38,4 @@ pub struct DagList {
 #[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Tag {
     pub name: String,
-}
-
-// From trait implementations for v1 models
-impl From<v1::model::dag::DagResponse> for Dag {
-    fn from(value: v1::model::dag::DagResponse) -> Self {
-        Self {
-            dag_id: value.dag_id.into(),
-            dag_display_name: Some(value.dag_display_name),
-            description: value.description,
-            fileloc: value.fileloc,
-            is_paused: value.is_paused.unwrap_or(false),
-            is_active: value.is_active,
-            has_import_errors: value.has_import_errors.unwrap_or(false),
-            has_task_concurrency_limits: value.has_task_concurrency_limits.unwrap_or(false),
-            last_parsed_time: value.last_parsed_time,
-            last_expired: value.last_expired,
-            max_active_tasks: value.max_active_tasks.unwrap_or(0),
-            max_active_runs: value.max_active_runs,
-            next_dagrun_logical_date: value.next_dagrun,
-            next_dagrun_data_interval_start: value.next_dagrun_data_interval_start,
-            next_dagrun_data_interval_end: value.next_dagrun_data_interval_end,
-            next_dagrun_create_after: value.next_dagrun_create_after,
-            owners: value.owners.clone(),
-            tags: value
-                .tags
-                .unwrap_or_default()
-                .into_iter()
-                .map(std::convert::Into::into)
-                .collect(),
-            file_token: value.file_token.clone(),
-            timetable_description: value.timetable_description,
-        }
-    }
-}
-
-impl From<v1::model::dag::DagCollectionResponse> for DagList {
-    fn from(value: v1::model::dag::DagCollectionResponse) -> Self {
-        Self {
-            dags: value
-                .dags
-                .into_iter()
-                .map(std::convert::Into::into)
-                .collect(),
-            total_entries: value.total_entries,
-        }
-    }
-}
-
-impl From<v1::model::dag::DagTagResponse> for Tag {
-    fn from(value: v1::model::dag::DagTagResponse) -> Self {
-        Self { name: value.name }
-    }
-}
-
-// From trait implementations for v2 models
-impl From<v2::model::dag::Dag> for Dag {
-    fn from(value: v2::model::dag::Dag) -> Self {
-        Self {
-            dag_id: value.dag_id.into(),
-            dag_display_name: Some(value.dag_display_name),
-            description: value.description,
-            fileloc: value.fileloc,
-            is_paused: value.is_paused,
-            is_active: None,
-            has_import_errors: value.has_import_errors,
-            has_task_concurrency_limits: value.has_task_concurrency_limits,
-            last_parsed_time: value.last_parsed_time,
-            last_expired: value.last_expired,
-            max_active_tasks: value.max_active_tasks,
-            max_active_runs: value.max_active_runs,
-            next_dagrun_logical_date: value.next_dagrun_logical_date,
-            next_dagrun_create_after: value.next_dagrun_run_after,
-            next_dagrun_data_interval_start: value.next_dagrun_data_interval_start,
-            next_dagrun_data_interval_end: value.next_dagrun_data_interval_end,
-            owners: value.owners,
-            tags: value
-                .tags
-                .into_iter()
-                .map(std::convert::Into::into)
-                .collect(),
-            file_token: value.file_token,
-            timetable_description: value.timetable_description,
-        }
-    }
-}
-
-impl From<v2::model::dag::DagList> for DagList {
-    fn from(value: v2::model::dag::DagList) -> Self {
-        Self {
-            dags: value
-                .dags
-                .into_iter()
-                .map(std::convert::Into::into)
-                .collect(),
-            total_entries: value.total_entries,
-        }
-    }
-}
-
-impl From<v2::model::dag::Tag> for Tag {
-    fn from(value: v2::model::dag::Tag) -> Self {
-        Self { name: value.name }
-    }
 }
