+++ src/airflow/model/common/mod.rs
@@ -9,7 +9,7 @@ pub mod task;
 pub mod taskinstance;
 
 // Re-export common types for easier access
-pub use dag::{Dag, DagList};
+pub use dag::{Dag, DagList, Tag};
 #[allow(unused_imports)]
 pub use dagrun::{DagRun, DagRunList, DagRunState, RunType};
 pub use dagstats::{DagStatistic, DagStatsResponse};
