+++ src/airflow/model/common/task.rs
@@ -1,7 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-use crate::airflow::client::{v1, v2};
-
 /// Common Task model representing a task definition in a DAG
 #[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Task {
@@ -12,38 +10,4 @@ pub struct Task {
 #[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct TaskList {
     pub tasks: Vec<Task>,
-}
-
-impl From<v1::model::task::TaskResponse> for Task {
-    fn from(value: v1::model::task::TaskResponse) -> Self {
-        Self {
-            task_id: value.task_id,
-            downstream_task_ids: value.downstream_task_ids,
-        }
-    }
-}
-
-impl From<v1::model::task::TaskCollectionResponse> for TaskList {
-    fn from(value: v1::model::task::TaskCollectionResponse) -> Self {
-        Self {
-            tasks: value.tasks.into_iter().map(Into::into).collect(),
-        }
-    }
-}
-
-impl From<v2::model::task::TaskResponse> for Task {
-    fn from(value: v2::model::task::TaskResponse) -> Self {
-        Self {
-            task_id: value.task_id,
-            downstream_task_ids: value.downstream_task_ids,
-        }
-    }
-}
-
-impl From<v2::model::task::TaskCollectionResponse> for TaskList {
-    fn from(value: v2::model::task::TaskCollectionResponse) -> Self {
-        Self {
-            tasks: value.tasks.into_iter().map(Into::into).collect(),
-        }
-    }
 }
