+++ src/airflow/model/common/taskinstance.rs
@@ -1,7 +1,5 @@
 use std::fmt;
 
-use crate::airflow::client::v1;
-use crate::airflow::client::v2;
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
 
@@ -127,86 +125,3 @@ impl TimeBounded for TaskInstance {
     }
 }
 
-// From trait implementations for v1 models
-impl From<v1::model::taskinstance::TaskInstanceResponse> for TaskInstance {
-    fn from(value: v1::model::taskinstance::TaskInstanceResponse) -> Self {
-        Self {
-            task_id: value.task_id.into(),
-            dag_id: value.dag_id.into(),
-            dag_run_id: value.dag_run_id.into(),
-            logical_date: Some(value.execution_date),
-            start_date: value.start_date,
-            end_date: value.end_date,
-            duration: value.duration,
-            state: value.state.map(|s| TaskInstanceState::from(s.as_str())),
-            try_number: value.try_number,
-            max_tries: value.max_tries,
-            map_index: value.map_index,
-            hostname: Some(value.hostname),
-            unixname: Some(value.unixname),
-            pool: value.pool,
-            pool_slots: value.pool_slots,
-            queue: value.queue,
-            priority_weight: value.priority_weight,
-            operator: value.operator,
-            queued_when: value.queued_when,
-            pid: value.pid,
-            note: value.note,
-        }
-    }
-}
-
-impl From<v1::model::taskinstance::TaskInstanceCollectionResponse> for TaskInstanceList {
-    fn from(value: v1::model::taskinstance::TaskInstanceCollectionResponse) -> Self {
-        Self {
-            task_instances: value
-                .task_instances
-                .into_iter()
-                .map(std::convert::Into::into)
-                .collect(),
-            total_entries: value.total_entries,
-        }
-    }
-}
-
-// From trait implementations for v2 models
-impl From<v2::model::taskinstance::TaskInstance> for TaskInstance {
-    fn from(value: v2::model::taskinstance::TaskInstance) -> Self {
-        Self {
-            task_id: value.task_id.into(),
-            dag_id: value.dag_id.into(),
-            dag_run_id: value.dag_run_id.into(),
-            logical_date: value.logical_date,
-            start_date: value.start_date,
-            end_date: value.end_date,
-            duration: value.duration,
-            state: value.state.map(|s| TaskInstanceState::from(s.as_str())),
-            try_number: value.try_number,
-            max_tries: value.max_tries,
-            map_index: value.map_index,
-            hostname: value.hostname,
-            unixname: value.unixname,
-            pool: value.pool,
-            pool_slots: value.pool_slots,
-            queue: value.queue,
-            priority_weight: value.priority_weight,
-            operator: value.operator,
-            queued_when: value.queued_when,
-            pid: value.pid,
-            note: value.note,
-        }
-    }
-}
-
-impl From<v2::model::taskinstance::TaskInstanceList> for TaskInstanceList {
-    fn from(value: v2::model::taskinstance::TaskInstanceList) -> Self {
-        Self {
-            task_instances: value
-                .task_instances
-                .into_iter()
-                .map(std::convert::Into::into)
-                .collect(),
-            total_entries: value.total_entries,
-        }
-    }
-}
